backend/requirements.txt
fastapi
uvicorn
valkey
pydantic
pwdlib
pyjwt
python-dotenv
psycopg2-binary
